data validation

+++ Controllers/CommentController.cs
@@ -27,13 +27,18 @@ public class CommentController : ControllerBase
     [HttpGet]
     public async Task<IActionResult> GetAll()
     {
+        if(!ModelState.IsValid)
+            return BadRequest(ModelState);
         var comments = await _commentRepository.GetAllAsync();
         return Ok(comments);
     }
 
-    [HttpGet("{id}")]
+    [HttpGet("{id:int}")]
     public async Task<IActionResult> GetById([FromRoute] int id)
     {
+        if(!ModelState.IsValid)
+            return BadRequest(ModelState);
+        
         var comment = await _commentRepository.GetByIdAsync(id);
 
         if (comment == null)
@@ -45,8 +50,11 @@ public class CommentController : ControllerBase
     }
 
 
-    [HttpPost("{stockId}")]
+    [HttpPost("{stockId:int}")]
     public async Task<IActionResult> Creat([FromRoute] int stockId, [FromBody] CreateCommentRequestDto createCommentDto){
+        
+        if(!ModelState.IsValid)
+            return BadRequest(ModelState);
 
         if (!await _stockRepository.StockExists(stockId))
         {
@@ -59,10 +67,13 @@ public class CommentController : ControllerBase
     }
     
     
-    [HttpPut("{id}")]
-    [Route("{id}")]
+    [HttpPut("{id:int}")]
+    [Route("{id:int}")]
     public async Task<IActionResult> Update([FromRoute] int id, [FromBody] UpdateCommentDto updateCommentDto)
     {
+        if(!ModelState.IsValid)
+            return BadRequest(ModelState);
+        
         var commentModel = await _commentRepository.UpdateAsync(id , updateCommentDto.ToCommentsFromUpdateDto());
         if (commentModel == null)
         {
@@ -71,10 +82,12 @@ public class CommentController : ControllerBase
         return Ok(commentModel.ToCommentDto());
     }
 
-    [HttpDelete("{id}")]
-    [Route("{id}")]
+    [HttpDelete("{id:int}")]
+    [Route("{id:int}")]
     public async Task<IActionResult> Delete([FromRoute] int id)
     {
+        if(!ModelState.IsValid)
+            return BadRequest(ModelState);
         var commentModel = await _commentRepository.DeleteAsync(id);
         if (commentModel == null)
         {
